+++ src/Atoms/Icon/index.ts
@@ -1,6 +1,7 @@
 import { ArrowRight } from './components/ArrowRight';
 import { Attachement } from './components/Attachement';
 import { Bank } from './components/Bank';
+import { Bell } from './components/Bell';
 import { Bolt } from './components/Bolt';
 import { CheckMark } from './components/CheckMark';
 import { CheckMarkCircle } from './components/CheckMarkCircle';
@@ -8,8 +9,11 @@ import { ChevronDown } from './components/ChevronDown';
 import { ChevronUp } from './components/ChevronUp';
 import { Cross } from './components/Cross';
 import { CrossCircle } from './components/CrossCircle';
+import { DiagonalWithCircle } from './components/DiagonalWithCircle';
 import { DirectionDown } from './components/DirectionDown';
 import { DirectionUp } from './components/DirectionUp';
+import { EyeWithPlus } from './components/EyeWithPlus';
+import { EyeWithPlusActive } from './components/EyeWithPlusActive';
 import { File } from './components/File';
 import { Flag } from './components/Flag';
 import { Globe } from './components/Globe';
@@ -33,6 +37,7 @@ export default {
   ArrowRight,
   Attachement,
   Bank,
+  Bell,
   Bolt,
   CheckMark,
   CheckMarkCircle,
@@ -40,8 +45,11 @@ export default {
   ChevronUp,
   Cross,
   CrossCircle,
+  DiagonalWithCircle,
   DirectionDown,
   DirectionUp,
+  EyeWithPlus,
+  EyeWithPlusActive,
   File,
   Flag,
   Globe,